requirements.txt
django==4.2.*
psycopg2-binary
pgvector
python-dotenv
openai
requests
